v0.2

+++ UP 6 GUI/UP 6.1.cs
@@ -6,7 +6,7 @@ namespace UP_6_GUI
     public partial class Form1 : Form
     {
         private int N;
-        private double[] arr;
+        private int[] arr;
         public Form1()
         {
             InitializeComponent();
@@ -20,7 +20,7 @@ namespace UP_6_GUI
                 AnswerBox.Text = "В поле N введены недопустимые данные! Введите целое число больше 0!";
                 return;
             }
-            arr = new double[N];
+            arr = new int[N];
             try
             {
                 var digits = ArrayBox.Text.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
@@ -39,7 +39,7 @@ namespace UP_6_GUI
                 AnswerBox.Text += "Изначальный массив: ";
                 for (int i = 0; i < N; i++)
                 {
-                    arr[i] = double.Parse(digits[i]);
+                    arr[i] = int.Parse(digits[i]);
                     AnswerBox.Text += arr[i] + " ";
                 }
             }
